Accommodated changes suggested by the linter and ran formatter
